+++ mptrack/Ctrl_seq.cpp
@@ -380,7 +380,7 @@ BOOL COrderList::PreTranslateMessage(MSG *pMsg)
 		CInputHandler* ih = (CMainFrame::GetMainFrame())->GetInputHandler();
 
 		//Translate message manually
-		UINT nChar = pMsg->wParam;
+		UINT nChar = (UINT)pMsg->wParam;
 		UINT nRepCnt = LOWORD(pMsg->lParam);
 		UINT nFlags = HIWORD(pMsg->lParam);
 		KeyEventType kT = ih->GetKeyEventType(nFlags);
